feat: change hook name
